Add files via upload
This is a document containing 3 screenshots of my Trello Board for Sprint Week #1. The Trello Board assigns a card to each task my group had to complete for each of the 5 Sprint Week Projects. As we started working on a task, it was moved from the "To Do" to the "Doing" section of the board. When the task was complete, the corresponding card was moved to the "Done" section. Each card had a label indicating which project it was from, a projected date of completion, which group member was assigned the task, and a description of the task. Picture # 1 was taken on February 28, 2022, which was at the start of Sprint Week. Picture # 2 was taken on March 1, 2022, which was in the middle of Sprint Week. Picture # 3 was taken on March 4, 2022, which was the day my group completed all the work.