Add files via upload
This is the Python Program group 22A -1 completed for Project # 1 of Sprint Week # 1. This program is based on the claims processing system for employees who take work trips for the NL Chocolate Company. It is a menu based program, which presents the user with 5 options; 
1) Enter an Employee Travel Claim
2) Fun Interview Question
3) Cool Things with Strings and Dates
4) Graph Monthly Claim Totals
5) Quit Program
Each of the first 4 menu options written as functions, while the 5th ends the program. The program also includes 2 other functions at the start; a function to format dollar values, and a function to validate phone number inputs. 
The menu, as well as Options 1,  2, and 5, were written by Ian Collins. Options 3 and 4 were written by Kara Balsom.